feat: split account and signer
