+++ src/signer/index.ts
@@ -1,2 +1,2 @@
-export * from './default';
 export * from './interface';
+export * from './default';
