+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './lib';
 export * from './api';
+export * from './signer';
