+++ src/types/signer.ts
@@ -1,0 +1,7 @@
+import { BigNumberish } from '../utils/number';
+import { Invocation } from './lib';
+
+export interface TransactionDetails extends Omit<Invocation, 'nonce' | 'signature'> {
+  walletAddress: string;
+  nonce: BigNumberish;
+}
